fast_api/login_page/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.0,<3.0.0 # Updated Pydantic version for potential Python 3.13 compatibility
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
weasyprint==60.2
jinja2==3.1.3
python-dotenv==1.0.1
supabase>=2.3.1 # Added for Supabase integration
